Require timeseries of type uint64_t*

+++ include/inform/time_series.h
@@ -5,6 +5,6 @@
 
 #include <inform/entropy.h>
 
-entropy inform_active_info(int const *series, size_t n, int base, uint64_t k);
+entropy inform_active_info(uint64_t const *series, size_t n, int base, uint64_t k);
 
-entropy inform_active_info_ensemble(int const *series, size_t m, size_t n, int base, uint64_t k);
+entropy inform_active_info_ensemble(uint64_t const *series, size_t m, size_t n, int base, uint64_t k);
